+++ test/scripts/default-branch.sh
@@ -1,9 +1,10 @@
 #!/bin/sh
 
-# checkout local branch but without remote branch
+# repo with a custom default branch name (given as first argument, e.g. master or trunk)
+# and a feature branch checked out, no remote
 
 # setup new git repo
-git init --initial-branch=main
+git init --initial-branch=$1
 # needed for git run in CI context where none of this is set
 git config --local user.email "you@example.com"
 git config --local user.name "Your Name"
@@ -11,26 +12,12 @@ git config --local user.name "Your Name"
 echo "hello" > file
 git add file
 git commit -a -m "initial commit"
-echo "hello" > file2
-git add .
-git commit -a -m "change 2"
-
-git checkout -b branch
-echo "hello2" > file
-git add file
-git commit -a -m "branch commit"
 
 git rev-parse HEAD > ../before.hash
 
-git checkout main
-
-git checkout -b branch2
-echo "another hello" > file3
-git add file3
-git commit -a -m "branch commit 2"
+git checkout -b branch
+echo "hello2" > file
+git commit -a -m "branch commit"
 
 # need the commit hash in the js code, and it's different each run
 git rev-parse HEAD > ../commit.hash
-
-# leave HEAD on main so that neither branch nor branch2 is the current branch
-git checkout main