+++ test/scripts/identical-branch.sh
@@ -1,0 +1,13 @@
+#!/bin/sh
+
+# single main branch, nothing else to compare against
+
+# setup new git repo
+git init --initial-branch=main
+# needed for git run in CI context where none of this is set
+git config --local user.email "you@example.com"
+git config --local user.name "Your Name"
+
+echo "hello" > file
+git add file
+git commit -a -m "initial commit"